agents/requirements.txt

fastapi>=0.68.0
uvicorn>=0.15.0
pydantic==2.5.0
google-genai>=0.1.0
faker==20.1.0
pandas==2.1.4
numpy==1.25.2
python-multipart==0.0.6
python-dotenv>=0.19.0